# Remove extra workspace queries in matrix inverse computation