@vishwakftw 2019-05-24
Speedup summary on batches of 3 x 3 matrices:

| Dimensions | Before this PR | After this PR | Improvement factor (Higher is better) |
| --- | --- | --- | --- |
| 100 x 3 x 3 | 95 µs ± 539 ns per loop | 24.7 µs ± 326 ns per loop | 3.85 |
| 200 x 3 x 3 | 185 µs ± 1.21 µs per loop | 43.6 µs ± 227 ns per loop | 4.23 |
| 500 x 3 x 3 | 450 µs ± 4.6 µs per loop | 99.1 µs ± 189 ns per loop | 4.54 |
| 1000 x 3 x 3 | 889 µs ± 4.5 µs per loop | 193 µs ± 586 ns per loop | 4.61 |
| 2000 x 3 x 3 | 1.74 ms ± 2.83 µs per loop | 379 µs ± 1.43 µs per loop | 4.59 |